requirements.txt
pydantic>=2.0.0
requests>=2.31.0
rich>=13.0.0
pypdf>=4.0.0
python-docx>=1.0.0
openpyxl>=3.0.0
pywin32; sys_platform == 'win32'
fastapi